Edited testing structure
