+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "java.project.sourcePaths": [],
+    "java.configuration.updateBuildConfiguration": "automatic"
+}
