--- evictionEducation/evictionEduspring/src/test/java/com/example/evictioneduspring/TestRunner.java
@@ -1,17 +1,0 @@
-package com.example.evictioneduspring;
-
-import org.junit.runner.RunWith;
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-
-
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        features = "src/test/resources/features/FrontendFeatures",  //Path to your feature files
-        glue = "com/example/evictioneduspring/stepdefinitions", // Package containing step definitions
-        plugin = {"pretty","html:target/cucumber-reports.html"}, //Report Generation
-        monochrome = true // Makes Output Readable
-)
-
-public class TestRunner {
-}
