--- evictionEducation/evictionEduspring/src/test/resources/features/DatabaseQueryFeatures/ExampleDatabaseQuery.feature
@@ -1,5 +1,0 @@
-@connectToH2
-Feature: Example Database Query
-
-  Scenario: I check that the first name of user 1 is Nick
-    Then the first name of user "1" is "Nick"
