+++ evictionEducation/evictionTests/src/test/java/com/example/evictiontests/constants/XPathConstants.java
@@ -1,0 +1,28 @@
+package com.example.evictiontests.constants;
+
+public class XPathConstants {
+
+    public static final String BACK_BUTTON = "//button[contains(text(), 'Back')]";
+
+    public static final String CHAT_NOW_BUTTON = "//a[contains(text(), 'Chat Now')]";
+    public static final String CONGRATS_TEXT = "//p[contains(text(), '🎉 Congrats, but additional resources may be available.')]";
+
+    public static final String FAQS_BUTTON = "//a[contains(text(), 'FAQs')]";
+    public static final String FIND_A_SOLUTION_BUTTON = "//a[contains(text(), 'Find A Solution')]";
+
+    public static final String I_AGREE_BUTTON = "//button[contains(text(), 'I agree')]";
+
+    public static final String NO_BUTTON = "//button[contains(text(), 'No')]";
+
+    public static final String OPEN_CHATBOT_BUTTON = "//a[contains(text(), 'Open Chatbot')]";
+    public static final String OPEN_MESSENGER_CHAT_BUTTON = "//button[contains(text(), 'Open Messenger Chat')]";
+
+
+    public static final String RESOURCES_BUTTON = "//a[contains(text(), 'Resources')]";
+
+    public static final String THIRTY_DAY_NOTICE_BUTTON = "//button[contains(text(), '30 Day Notice')]";
+    public static final String THREE_DAY_NOTICE_BUTTON = "//button[contains(text(), '3 Day Notice')]";
+
+    public static final String YES_BUTTON = "//button[contains(text(), 'Yes')]";
+
+}
